+++ snowddl/__init__.py
@@ -3,6 +3,7 @@ from .blueprint import *
 from .config import SnowDDLConfig
 from .engine import SnowDDLEngine
 from .formatter import SnowDDLFormatter
+from .query_builder import SnowDDLQueryBuilder
 from .settings import SnowDDLSettings
 from .error import SnowDDLExecuteError, SnowDDLUnsupportedError
 from .parser import *